Bump release to 1.4.0
TODO: Fix upgrading w/ ansible. Currently when you bump the release in the
ansible config it'll download and configure the new release, but not update the
symlink to the `latest` directory and restart the mattermost systemd unit.

TODO: Fix a JSON whoopsie in the config.json which prevents mattermost from
starting.